# schema/openapi/openapi.yaml
openapi: '3.0.3'
info:
  title: Backbone API
  version: '1.0.0'
  description: |
    Auto-generated by backbone-cli.
    This API provides 11 standard CRUD endpoints for each model.
servers:
  - url: http://localhost:3000
    description: Development server
...
paths:
...
  /api/v1/registrations:
    get:
      tags:
        - Registrations
      summary: List Registrations
      description: Get a paginated list of Registrations with optional filtering and sorting
      operationId: listRegistrations
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegistrationList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - Registrations
      summary: Create Registration
      description: Create a new Registration
      operationId: createRegistration
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRegistrationRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Registration'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  schemas:
...
    Registration:
      type: object
      required:
        - id
        - event_id
        - name
        - email
        - state
        - active
        - barcode
        - metadata
      properties:
        id:
          type: string
          format: uuid
        event_id:
          type: string
          format: uuid
        event_slot_id:
          nullable: true
          type: string
          format: uuid
        event_ticket_id:
          nullable: true
          type: string
          format: uuid
        name:
          type: string
          maxLength: 200
        email:
          type: string
          format: email
          maxLength: 254
        phone:
          nullable: true
          type: string
          maxLength: 40
        company_name:
          nullable: true
          type: string
          maxLength: 200
        partner_id:
          nullable: true
          type: string
          format: uuid
        state:
          type: string
        date_closed:
          nullable: true
          type: string
          format: date-time
        sale_order_id:
          nullable: true
          type: string
          format: uuid
        sale_order_state:
          nullable: true
          type: string
        sale_status:
          nullable: true
          type: string
        active:
          type: boolean
        barcode:
          type: string
          maxLength: 32
        company_id:
          nullable: true
          type: string
          format: uuid
        metadata:
          type: object
...
    CreateRegistrationRequest:
      type: object
      required:
        - event_id
        - name
        - email
        - state
        - active
        - barcode
        - metadata
      properties:
        event_id:
          type: string
          format: uuid
        event_slot_id:
          nullable: true
          type: string
          format: uuid
        event_ticket_id:
          nullable: true
          type: string
          format: uuid
        name:
          type: string
          maxLength: 200
        email:
          type: string
          format: email
          maxLength: 254
        phone:
          nullable: true
          type: string
          maxLength: 40
        company_name:
          nullable: true
          type: string
          maxLength: 200
        partner_id:
          nullable: true
          type: string
          format: uuid
        state:
          type: string
        date_closed:
          nullable: true
          type: string
          format: date-time
        sale_order_id:
          nullable: true
          type: string
          format: uuid
        sale_order_state:
          nullable: true
          type: string
        sale_status:
          nullable: true
          type: string
        active:
          type: boolean
        barcode:
          type: string
          maxLength: 32
        company_id:
          nullable: true
          type: string
          format: uuid
        metadata:
          type: object
...
    RegistrationList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Registration'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
...
    PaginationMeta:
      type: object
      properties:
        total:
          type: integer
          example: 100
        page:
          type: integer
          example: 1
        per_page:
          type: integer
          example: 20
        total_pages:
          type: integer
          example: 5
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
          example: Invalid input data
        details:
          type: object
          additionalProperties: true
  parameters:
...
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PerPageParam:
      name: per_page
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    SortByParam:
      name: sort_by
      in: query
      schema:
        type: string
    SortOrderParam:
      name: sort_order
      in: query
      schema:
        type: string
        enum: [asc, desc]
        default: asc
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []