# src/presentation/http/registration_handler.rs
//! Registration REST handlers
//!
//! Generated by metaphor-schema. Do not edit manually.
//!
//! Uses Axum and backbone-core's BackboneCrudHandler for all 12 CRUD endpoints.

use std::collections::HashMap;
use std::sync::Arc;

use axum::Router;
use serde::{Deserialize, Serialize};
use uuid::Uuid;
use chrono::{DateTime, Utc};

// Backbone framework imports
use backbone_core::http::{ApiResponse, BackboneCrudHandler};

// Auth integration (optional)
#[cfg(feature = "auth")]
use backbone_auth::middleware::AuthContext;
#[cfg(feature = "auth")]
use backbone_auth::AuthMiddleware;

// Domain imports
use crate::domain::entity::*;
use crate::application::service::{RegistrationService, ServiceError};

// DTO imports
use crate::presentation::dto::{CreateRegistrationDto, UpdateRegistrationDto, PatchRegistrationDto, RegistrationResponseDto};

use crate::domain::state_machine::{registration_stateState, registration_stateStateMachine, registration_stateTransition};

/// Application error type
#[derive(Debug, thiserror::Error)]
pub enum RegistrationError {
    #[error("Not found: {0}")]
    NotFound(String),
    #[error("Validation error: {0}")]
    Validation(String),
    #[error("Database error: {0}")]
    Database(String),
    #[error("Internal error: {0}")]
    Internal(String),
}

impl From<ServiceError> for RegistrationError {
    fn from(err: ServiceError) -> Self {
        match err {
            ServiceError::NotFound => Self::NotFound(err.to_string()),
            ServiceError::Validation(ref msg) => Self::Validation(msg.clone()),
            ServiceError::AlreadyExists(ref msg) => Self::Validation(msg.clone()),
            ServiceError::Repository(ref e) => Self::Database(e.to_string()),
            ServiceError::Internal(ref msg) => Self::Internal(msg.clone()),
        }
    }
}

impl axum::response::IntoResponse for RegistrationError {
    fn into_response(self) -> axum::response::Response {
        use axum::http::StatusCode;
        use axum::Json;

        let (status, code) = match &self {
            Self::NotFound(_) => (StatusCode::NOT_FOUND, "REGISTRATION_NOT_FOUND"),
            Self::Validation(_) => (StatusCode::BAD_REQUEST, "REGISTRATION_VALIDATION_ERROR"),
            Self::Database(_) => (StatusCode::INTERNAL_SERVER_ERROR, "REGISTRATION_DATABASE_ERROR"),
            Self::Internal(_) => (StatusCode::INTERNAL_SERVER_ERROR, "REGISTRATION_INTERNAL_ERROR"),
        };

        let body = serde_json::json!({
            "success": false,
            "error": code,
            "message": self.to_string(),
        });

        (status, Json(body)).into_response()
    }
}

// =============================================================================
// Route Configuration
// =============================================================================

/// Create Axum router with all 16 Backbone endpoints for Registration.
///
/// # Routes
///
/// | Method | Path | Description |
/// |--------|------|-------------|
/// | GET | /registrations | List with pagination |
/// | POST | /registrations | Create new |
/// | GET | /registrations/:id | Get by ID |
/// | PUT | /registrations/:id | Full update |
/// | PATCH | /registrations/:id | Partial update |
/// | DELETE | /registrations/:id | Soft delete |
/// | POST | /registrations/bulk | Bulk create |
/// | POST | /registrations/upsert | Upsert |
/// | GET | /registrations/trash | List deleted |
/// | POST | /registrations/:id/restore | Restore |
/// | DELETE | /registrations/empty | Empty trash |
/// | GET | /registrations/:id/deleted | Get deleted by ID |
/// | DELETE | /registrations/trash/:id | Permanent delete from trash |
/// | GET | /registrations/count | Count active entities |
/// | GET | /registrations/trash/count | Count deleted entities |
///
/// # Example
///
/// ```text
/// let service = Arc::new(RegistrationService::with_repository(repository));
/// let router = create_registration_routes(service);
/// ```
pub fn create_registration_routes(service: Arc<RegistrationService>) -> Router {
    BackboneCrudHandler::<RegistrationService, Registration, CreateRegistrationDto, UpdateRegistrationDto, RegistrationResponseDto>::routes(
        service,
        "/registrations",
    )
}

/// Create Axum router with only the read (GET) endpoints for Registration.
///
/// Safe for public, unauthenticated exposure (e.g., reference data).
/// Mutations must be served separately via `create_registration_write_routes`,
/// typically wrapped in an auth middleware layer.
pub fn create_registration_read_routes(service: Arc<RegistrationService>) -> Router {
    BackboneCrudHandler::<RegistrationService, Registration, CreateRegistrationDto, UpdateRegistrationDto, RegistrationResponseDto>::read_routes(
        service,
        "/registrations",
    )
}

/// Create Axum router with only the write (mutation) endpoints for Registration.
///
/// These routes must NOT be publicly exposed. Wrap them with an auth
/// middleware before nesting into the application router.
///
/// # This is unguarded generic CRUD, not a validated write path
///
/// These are plain create/update/patch/delete mutations over the entity row —
/// they bypass all business invariants. If the module exposes a validated write
/// service (e.g. a command router over its domain engine), serve THAT instead
/// for any mutation that must respect domain rules.
pub fn create_registration_write_routes(service: Arc<RegistrationService>) -> Router {
    BackboneCrudHandler::<RegistrationService, Registration, CreateRegistrationDto, UpdateRegistrationDto, RegistrationResponseDto>::write_routes(
        service,
        "/registrations",
    )
}

/// Create authenticated routes with auth middleware.
///
/// Requires the `auth` feature flag. The `AuthMiddleware` implementation
/// is responsible for extracting and validating tokens, then providing
/// an `AuthContext` via request extensions.
#[cfg(feature = "auth")]
pub fn create_protected_registration_routes<A: AuthMiddleware + Send + Sync + 'static>(
    service: Arc<RegistrationService>,
    auth: Arc<A>,
) -> Router {
    use axum::middleware;
    use axum::response::IntoResponse;

    let auth_layer = auth.clone();
    create_registration_routes(service)
        .layer(middleware::from_fn(move |mut req: axum::extract::Request, next: axum::middleware::Next| {
            let auth = auth_layer.clone();
            async move {
                let token = req.headers()
                    .get(axum::http::header::AUTHORIZATION)
                    .and_then(|h| h.to_str().ok())
                    .and_then(|raw| raw.strip_prefix("Bearer ").or_else(|| raw.strip_prefix("bearer ")))
                    .unwrap_or("");
                match auth.authenticate(token).await {
                    Ok(ctx) => {
                        req.extensions_mut().insert(ctx);
                        next.run(req).await
                    }
                    Err(_) => {
                        (axum::http::StatusCode::UNAUTHORIZED,
                         axum::Json(serde_json::json!({
                             "success": false,
                             "error": "unauthorized",
                             "message": "Authentication required"
                         }))
                        ).into_response()
                    }
                }
            }
        }))
}

// =============================================================================
// State Transition Handlers
// =============================================================================

/// Execute set_draft transition on a Registration.
///
/// POST /registrations/:id/transitions/set_draft
pub async fn set_draft_transition(
    axum::extract::State(service): axum::extract::State<Arc<RegistrationService>>,
    axum::extract::Path(id): axum::extract::Path<String>,
    #[cfg(feature = "auth")] axum::Extension(auth): axum::Extension<AuthContext>,
) -> impl axum::response::IntoResponse {
    use axum::{http::StatusCode, Json};

    // Get current entity
    let entity = match service.get_by_id(&id).await {
        Ok(Some(e)) => e,
        Ok(None) => return (StatusCode::NOT_FOUND, Json(ApiResponse::<RegistrationResponseDto>::not_found("Registration", &id))),
        Err(e) => return (StatusCode::INTERNAL_SERVER_ERROR, Json(ApiResponse::<RegistrationResponseDto>::error(e.to_string()))),
    };

    // Check permission (if auth enabled)
    #[cfg(feature = "auth")]
    {
        let allowed_roles = registration_stateTransition::SetDraft.allowed_roles();
        let has_specific_perm = auth.permissions.iter().any(|p| p == "registration:transition:set_draft");
        let has_update_perm = auth.permissions.iter().any(|p| p == "registration:update");
        if !has_specific_perm && !has_update_perm {
            return (StatusCode::FORBIDDEN, Json(ApiResponse::<RegistrationResponseDto>::error("Insufficient permissions for set_draft transition")));
        }
    }

    // Create state machine from entity's actual status and validate transition
    let current_state: registration_stateState = entity.state.to_string().parse()
        .unwrap_or(registration_stateState::default());
    let sm = registration_stateStateMachine::from_state(current_state);
    if !sm.can_transition(registration_stateTransition::SetDraft) {
        return (StatusCode::BAD_REQUEST, Json(ApiResponse::<RegistrationResponseDto>::error("Transition not allowed from current state")));
    }

    // Apply transition via partial update
    let mut fields: HashMap<String, serde_json::Value> = HashMap::new();
    fields.insert("state".to_string(), serde_json::Value::String("Draft".to_string()));

    match service.partial_update(&id, fields).await {
        Ok(Some(updated)) => {
            let response: RegistrationResponseDto = updated.into();
            (StatusCode::OK, Json(ApiResponse::ok(response)))
        }
        Ok(None) => (StatusCode::NOT_FOUND, Json(ApiResponse::<RegistrationResponseDto>::not_found("Registration", &id))),
        Err(e) => (StatusCode::INTERNAL_SERVER_ERROR, Json(ApiResponse::<RegistrationResponseDto>::error(e.to_string()))),
    }
}

/// Execute confirm transition on a Registration.
///
/// POST /registrations/:id/transitions/confirm
pub async fn confirm_transition(
    axum::extract::State(service): axum::extract::State<Arc<RegistrationService>>,
    axum::extract::Path(id): axum::extract::Path<String>,
    #[cfg(feature = "auth")] axum::Extension(auth): axum::Extension<AuthContext>,
) -> impl axum::response::IntoResponse {
    use axum::{http::StatusCode, Json};

    // Get current entity
    let entity = match service.get_by_id(&id).await {
        Ok(Some(e)) => e,
        Ok(None) => return (StatusCode::NOT_FOUND, Json(ApiResponse::<RegistrationResponseDto>::not_found("Registration", &id))),
        Err(e) => return (StatusCode::INTERNAL_SERVER_ERROR, Json(ApiResponse::<RegistrationResponseDto>::error(e.to_string()))),
    };

    // Check permission (if auth enabled)
    #[cfg(feature = "auth")]
    {
        let allowed_roles = registration_stateTransition::Confirm.allowed_roles();
        let has_specific_perm = auth.permissions.iter().any(|p| p == "registration:transition:confirm");
        let has_update_perm = auth.permissions.iter().any(|p| p == "registration:update");
        if !has_specific_perm && !has_update_perm {
            return (StatusCode::FORBIDDEN, Json(ApiResponse::<RegistrationResponseDto>::error("Insufficient permissions for confirm transition")));
        }
    }

    // Create state machine from entity's actual status and validate transition
    let current_state: registration_stateState = entity.state.to_string().parse()
        .unwrap_or(registration_stateState::default());
    let sm = registration_stateStateMachine::from_state(current_state);
    if !sm.can_transition(registration_stateTransition::Confirm) {
        return (StatusCode::BAD_REQUEST, Json(ApiResponse::<RegistrationResponseDto>::error("Transition not allowed from current state")));
    }

    // Apply transition via partial update
    let mut fields: HashMap<String, serde_json::Value> = HashMap::new();
    fields.insert("state".to_string(), serde_json::Value::String("Open".to_string()));

    match service.partial_update(&id, fields).await {
        Ok(Some(updated)) => {
            let response: RegistrationResponseDto = updated.into();
            (StatusCode::OK, Json(ApiResponse::ok(response)))
        }
        Ok(None) => (StatusCode::NOT_FOUND, Json(ApiResponse::<RegistrationResponseDto>::not_found("Registration", &id))),
        Err(e) => (StatusCode::INTERNAL_SERVER_ERROR, Json(ApiResponse::<RegistrationResponseDto>::error(e.to_string()))),
    }
}

/// Execute set_done transition on a Registration.
///
/// POST /registrations/:id/transitions/set_done
pub async fn set_done_transition(
    axum::extract::State(service): axum::extract::State<Arc<RegistrationService>>,
    axum::extract::Path(id): axum::extract::Path<String>,
    #[cfg(feature = "auth")] axum::Extension(auth): axum::Extension<AuthContext>,
) -> impl axum::response::IntoResponse {
    use axum::{http::StatusCode, Json};

    // Get current entity
    let entity = match service.get_by_id(&id).await {
        Ok(Some(e)) => e,
        Ok(None) => return (StatusCode::NOT_FOUND, Json(ApiResponse::<RegistrationResponseDto>::not_found("Registration", &id))),
        Err(e) => return (StatusCode::INTERNAL_SERVER_ERROR, Json(ApiResponse::<RegistrationResponseDto>::error(e.to_string()))),
    };

    // Check permission (if auth enabled)
    #[cfg(feature = "auth")]
    {
        let allowed_roles = registration_stateTransition::SetDone.allowed_roles();
        let has_specific_perm = auth.permissions.iter().any(|p| p == "registration:transition:set_done");
        let has_update_perm = auth.permissions.iter().any(|p| p == "registration:update");
        if !has_specific_perm && !has_update_perm {
            return (StatusCode::FORBIDDEN, Json(ApiResponse::<RegistrationResponseDto>::error("Insufficient permissions for set_done transition")));
        }
    }

    // Create state machine from entity's actual status and validate transition
    let current_state: registration_stateState = entity.state.to_string().parse()
        .unwrap_or(registration_stateState::default());
    let sm = registration_stateStateMachine::from_state(current_state);
    if !sm.can_transition(registration_stateTransition::SetDone) {
        return (StatusCode::BAD_REQUEST, Json(ApiResponse::<RegistrationResponseDto>::error("Transition not allowed from current state")));
    }

    // Apply transition via partial update
    let mut fields: HashMap<String, serde_json::Value> = HashMap::new();
    fields.insert("state".to_string(), serde_json::Value::String("Done".to_string()));

    match service.partial_update(&id, fields).await {
        Ok(Some(updated)) => {
            let response: RegistrationResponseDto = updated.into();
            (StatusCode::OK, Json(ApiResponse::ok(response)))
        }
        Ok(None) => (StatusCode::NOT_FOUND, Json(ApiResponse::<RegistrationResponseDto>::not_found("Registration", &id))),
        Err(e) => (StatusCode::INTERNAL_SERVER_ERROR, Json(ApiResponse::<RegistrationResponseDto>::error(e.to_string()))),
    }
}

/// Execute cancel transition on a Registration.
///
/// POST /registrations/:id/transitions/cancel
pub async fn cancel_transition(
    axum::extract::State(service): axum::extract::State<Arc<RegistrationService>>,
    axum::extract::Path(id): axum::extract::Path<String>,
    #[cfg(feature = "auth")] axum::Extension(auth): axum::Extension<AuthContext>,
) -> impl axum::response::IntoResponse {
    use axum::{http::StatusCode, Json};

    // Get current entity
    let entity = match service.get_by_id(&id).await {
        Ok(Some(e)) => e,
        Ok(None) => return (StatusCode::NOT_FOUND, Json(ApiResponse::<RegistrationResponseDto>::not_found("Registration", &id))),
        Err(e) => return (StatusCode::INTERNAL_SERVER_ERROR, Json(ApiResponse::<RegistrationResponseDto>::error(e.to_string()))),
    };

    // Check permission (if auth enabled)
    #[cfg(feature = "auth")]
    {
        let allowed_roles = registration_stateTransition::Cancel.allowed_roles();
        let has_specific_perm = auth.permissions.iter().any(|p| p == "registration:transition:cancel");
        let has_update_perm = auth.permissions.iter().any(|p| p == "registration:update");
        if !has_specific_perm && !has_update_perm {
            return (StatusCode::FORBIDDEN, Json(ApiResponse::<RegistrationResponseDto>::error("Insufficient permissions for cancel transition")));
        }
    }

    // Create state machine from entity's actual status and validate transition
    let current_state: registration_stateState = entity.state.to_string().parse()
        .unwrap_or(registration_stateState::default());
    let sm = registration_stateStateMachine::from_state(current_state);
    if !sm.can_transition(registration_stateTransition::Cancel) {
        return (StatusCode::BAD_REQUEST, Json(ApiResponse::<RegistrationResponseDto>::error("Transition not allowed from current state")));
    }

    // Apply transition via partial update
    let mut fields: HashMap<String, serde_json::Value> = HashMap::new();
    fields.insert("state".to_string(), serde_json::Value::String("Cancel".to_string()));

    match service.partial_update(&id, fields).await {
        Ok(Some(updated)) => {
            let response: RegistrationResponseDto = updated.into();
            (StatusCode::OK, Json(ApiResponse::ok(response)))
        }
        Ok(None) => (StatusCode::NOT_FOUND, Json(ApiResponse::<RegistrationResponseDto>::not_found("Registration", &id))),
        Err(e) => (StatusCode::INTERNAL_SERVER_ERROR, Json(ApiResponse::<RegistrationResponseDto>::error(e.to_string()))),
    }
}

/// Create routes for state transitions.
pub fn create_registration_transition_routes(service: Arc<RegistrationService>) -> Router {
    use axum::routing::post;

    Router::new()
        .route("/registrations/:id/transitions/set_draft", post(set_draft_transition))
        .route("/registrations/:id/transitions/confirm", post(confirm_transition))
        .route("/registrations/:id/transitions/set_done", post(set_done_transition))
        .route("/registrations/:id/transitions/cancel", post(cancel_transition))
        .with_state(service)
}